M3: PDB backend (LLVM-backed); default in VSCode extension
Adds CodeView/PDB emission path alongside the existing DWARF one so
Delphi binaries can be debugged with VS native debugger (cppvsdbg) /
WinDbg without the gdb dependency.

Pipeline: rsm2pdb pdb <map> <exe-in-place> <pdb>
  - parses PE for sections, .map for publics/lines/modules, optional
    .rsm for proc params + locals
  - writes a PDB via llvm::pdb::PDBFileBuilder with: S_PUB32 publics,
    S_GDATA32 globals (unqualified names), per-module S_GPROC32 +
    S_FRAMEPROC + S_REGREL32 + S_END, C13 line subsections with MD5
    file checksums, SectionContribs sorted by (ISect, Off),
    DbgStream[SectionHdr] verbatim
  - injects an RSDS Debug Directory entry into the PE pointing at the
    PDB; patches DataDirectory[IMAGE_DIRECTORY_ENTRY_DEBUG]

VSCode extension gains rsm2pdb.backend (default "pdb", "dwarf" opt-in);
PDB mode launches cppvsdbg, DWARF mode keeps cppdbg + gdb. dprojParser
now expands $(Config) (not just $(Platform)) and resolves relative
DCC_ExeOutput against the .dproj dir -- needed for the AdvPcbTools
shape (Config=Debug/Release with no Cfg_N_Plat refinement).

End-to-end verified on examples/02_two_units: BPs in .dpr and .pas
hit, step-into enters Pascal source, Locals shows A/B/Tmp/Result,
Watch on S/D/P1 resolves globals (typed as Int32 raw view; real
Pascal TPI types are the next milestone).

Hard-won pitfalls now in CLAUDE.md gotchas 19-21 (reserved streams,
SectionHdr DbgStream, MD5 checksums, ProcSym.End linking,
SectionContrib sort, S_FRAMEPROC frame-ptr encoding, qualified
Pascal name parsing in cppvsdbg, source path canonicalization).

Tests: 31/31 pass (4 new in test_pdb_writer + test_pe_pdb_injector).

+++ examples/02_two_units/two_units.dpr
@@ -15,7 +15,7 @@ uses
 var
   P1, P2: Geometry.TPoint;
   C:      App.Colors.TColor;
-  D:      Integer;
+  D:      Double;
   S:      Integer;
 
 begin
